Google: rename Stackdriver to Monitoring
Keep the old Stackdriver modules as deprecated wrappers while moving the primary Google provider APIs to Cloud Monitoring and Cloud Logging.
