+++ providers/google/provider.yaml
@@ -261,8 +261,8 @@ integrations:
       - /docs/apache-airflow-providers-google/operators/cloud/cloud_sql.rst
     logo: /docs/integration-logos/Cloud-SQL.png
     tags: [gcp]
-  - integration-name: Google Cloud Stackdriver
-    external-doc-url: https://cloud.google.com/products/operations/
+  - integration-name: Google Cloud Monitoring
+    external-doc-url: https://cloud.google.com/monitoring/
     how-to-guide:
       - /docs/apache-airflow-providers-google/operators/cloud/stackdriver.rst
     logo: /docs/integration-logos/Google-Cloud-Stackdriver.png
@@ -554,8 +554,9 @@ operators:
   - integration-name: Google Cloud Speech-to-Text
     python-modules:
       - airflow.providers.google.cloud.operators.speech_to_text
-  - integration-name: Google Cloud Stackdriver
+  - integration-name: Google Cloud Monitoring
     python-modules:
+      - airflow.providers.google.cloud.operators.monitoring
       - airflow.providers.google.cloud.operators.stackdriver
   - integration-name: Google Cloud Tasks
     python-modules:
@@ -829,8 +830,9 @@ hooks:
   - integration-name: Google Cloud Speech-to-Text
     python-modules:
       - airflow.providers.google.cloud.hooks.speech_to_text
-  - integration-name: Google Cloud Stackdriver
+  - integration-name: Google Cloud Monitoring
     python-modules:
+      - airflow.providers.google.cloud.hooks.monitoring
       - airflow.providers.google.cloud.hooks.stackdriver
   - integration-name: Google Cloud Tasks
     python-modules:
@@ -1407,8 +1409,8 @@ extra-links:
   - airflow.providers.google.cloud.links.bigtable.BigtableTablesLink
   - airflow.providers.google.cloud.links.spanner.SpannerDatabaseLink
   - airflow.providers.google.cloud.links.spanner.SpannerInstanceLink
-  - airflow.providers.google.cloud.links.stackdriver.StackdriverNotificationsLink
-  - airflow.providers.google.cloud.links.stackdriver.StackdriverPoliciesLink
+  - airflow.providers.google.cloud.links.monitoring.CloudMonitoringNotificationsLink
+  - airflow.providers.google.cloud.links.monitoring.CloudMonitoringPoliciesLink
   - airflow.providers.google.cloud.links.kubernetes_engine.KubernetesEngineClusterLink
   - airflow.providers.google.cloud.links.kubernetes_engine.KubernetesEnginePodLink
   - airflow.providers.google.cloud.links.kubernetes_engine.KubernetesEngineJobLink
@@ -1468,7 +1470,7 @@ auth-backends:
 
 logging:
   - airflow.providers.google.cloud.log.gcs_task_handler.GCSTaskHandler
-  - airflow.providers.google.cloud.log.stackdriver_task_handler.StackdriverTaskHandler
+  - airflow.providers.google.cloud.log.cloud_logging_task_handler.CloudLoggingTaskHandler
 
 queues:
   - airflow.providers.google.event_scheduling.events.pubsub.PubSubMessageQueueEventTriggerContainer
