+++ providers/google/src/airflow/providers/google/cloud/links/monitoring.py
@@ -1,0 +1,42 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+"""This module contains Google Cloud Monitoring links."""
+
+from __future__ import annotations
+
+from airflow.providers.google.cloud.links.base import BaseGoogleLink
+
+MONITORING_BASE_LINK = "/monitoring/alerting"
+CLOUD_MONITORING_NOTIFICATIONS_LINK = MONITORING_BASE_LINK + "/notifications?project={project_id}"
+CLOUD_MONITORING_POLICIES_LINK = MONITORING_BASE_LINK + "/policies?project={project_id}"
+
+
+class CloudMonitoringNotificationsLink(BaseGoogleLink):
+    """Helper class for constructing Cloud Monitoring Notifications Link."""
+
+    name = "Cloud Monitoring Notifications"
+    key = "cloud_monitoring_notifications"
+    format_str = CLOUD_MONITORING_NOTIFICATIONS_LINK
+
+
+class CloudMonitoringPoliciesLink(BaseGoogleLink):
+    """Helper class for constructing Cloud Monitoring Policies Link."""
+
+    name = "Cloud Monitoring Policies"
+    key = "cloud_monitoring_policies"
+    format_str = CLOUD_MONITORING_POLICIES_LINK
